% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runPipeline.R
\name{runPipeline}
\alias{runPipeline}
\title{runPipeline}
\usage{
runPipeline(
  workdir = c(),
  datadir = c(),
  configfile = c(),
  verbose = TRUE,
  timetag = TRUE
)
}
\arguments{
\item{workdir}{path to location where code can store test files and results}

\item{datadir}{path to directory where the accelerometer files to test are stored.}

\item{configfile}{path to GGIR configuration file.}

\item{verbose}{verbose}

\item{timetag}{Boolean to indicate whether timestamp should be added to filenames}
}
\value{
Stores the GGIR output based on the configuration file provided and compares
it with the previous output stored in the same workdir.
}
\description{
runPipeline
}
